requirements.txt
flask>=2.3.0
python-dotenv>=1.0.0
twilio>=8.0.0
openai>=1.12.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyPDF2>=3.0.0
Pillow>=10.0.0
gunicorn>=21.0.0
werkzeug>=2.3.0